dom/workers/test/serviceworkers/test_privateBrowsing.html

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
  <title>Test for ServiceWorker - Private Browsing</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
</head>
<body>

<script type="application/javascript">

const Ci = Components.interfaces;
var mainWindow;

var contentPage = "http://mochi.test:8888/chrome/dom/workers/test/empty.html";

function testOnWindow(aIsPrivate, aCallback) {
  var win = mainWindow.OpenBrowserWindow({private: aIsPrivate});
  win.addEventListener("load", function onLoad() {
    win.removeEventListener("load", onLoad, false);
    win.addEventListener("DOMContentLoaded", function onInnerLoad() {
      if (win.content.location.href != contentPage) {
        win.gBrowser.loadURI(contentPage);
        return;
      }

      win.removeEventListener("DOMContentLoaded", onInnerLoad, true);
      SimpleTest.executeSoon(function() { aCallback(win); });
    }, true);

    if (!aIsPrivate) {
      win.gBrowser.loadURI(contentPage);
    }
  }, true);
}

function setupWindow() {
  mainWindow = window.QueryInterface(Ci.nsIInterfaceRequestor)
                     .getInterface(Ci.nsIWebNavigation)
                     .QueryInterface(Ci.nsIDocShellTreeItem)
                     .rootTreeItem
                     .QueryInterface(Ci.nsIInterfaceRequestor)
                     .getInterface(Ci.nsIDOMWindow);
  runTest();
}

var wN;
var wP;

function doTests() {
  testOnWindow(false, function(aWin) {
    wN = aWin;
    ok("serviceWorker" in wN.content.navigator, "ServiceWorkers are available for normal windows");

    testOnWindow(true, function(aWin) {
      wP = aWin;
      ok(!("serviceWorker" in wP.content.navigator), "ServiceWorkers are not available for private windows");
      runTest();
    });
  });
}

var steps = [
  setupWindow,
  doTests
];

function runTest() {
  if (!steps.length) {
    wN.close();
    wP.close();

    SimpleTest.finish();
    return;
  }

  var step = steps.shift();
  step();
}

SimpleTest.waitForExplicitFinish();
SpecialPowers.pushPrefEnv({"set": [
  ["dom.serviceWorkers.enabled", true],
  ["browser.startup.page", 0],
  ["browser.startup.homepage_override.mstone", "ignore"],
]}, runTest);

</script>
</body>
</html>
